:muscle: Individually define utility functions
