+++ assert.ts
@@ -92,49 +92,3 @@ export function assert<T>(
     );
   }
 }
-
-/**
- * Ensures that the given value satisfies the provided predicate.
- *
- * ```ts
- * import { ensure, is } from "@core/unknownutil";
- *
- * const a: unknown = "hello";
- * const _: string = ensure(a, is.String);
- * ```
- *
- * @param x The value to be ensured.
- * @param pred The predicate function to test the value against.
- * @param options Optional configuration for the assertion.
- * @returns The input value `x`.
- * @throws {AssertError} If the value does not satisfy the predicate.
- */
-export function ensure<T>(
-  x: unknown,
-  pred: Predicate<T>,
-  options: { message?: string; name?: string } = {},
-): T {
-  assert(x, pred, options);
-  return x;
-}
-
-/**
- * Returns the input value if it satisfies the provided predicate, or `undefined` otherwise.
- *
- * ```ts
- * import { is, maybe } from "@core/unknownutil";
- *
- * const a: unknown = "hello";
- * const _: string = maybe(a, is.String) ?? "default value";
- * ```
- *
- * @param x The value to be tested.
- * @param pred The predicate function to test the value against.
- * @returns The input value `x` if it satisfies the predicate, or `undefined` otherwise.
- */
-export function maybe<T>(
-  x: unknown,
-  pred: Predicate<T>,
-): T | undefined {
-  return pred(x) ? x : undefined;
-}
