+++ mod.ts
@@ -241,6 +241,10 @@
  */
 
 export type * from "./type.ts";
+
 export * from "./as/mod.ts";
 export * from "./is/mod.ts";
-export * from "./util.ts";
+
+export * from "./assert.ts";
+export * from "./ensure.ts";
+export * from "./maybe.ts";
